requirements.txt
Django==4.1
psycopg2-binary==2.9.3
pandas==1.4.3
numpy==1.23.1
requests==2.28.1
celery==5.2.7
redis==4.2.2
python-decouple==3.6
reportlab==3.6.12
matplotlib==3.5.2
xgboost==2.1.1
lightgbm==3.3.2
optuna==3.0.3
scikit-learn==1.1.1
joblib==1.1.0
djangorestframework==3.13.1
django_celery_beat==2.5.0
django-celery-results==2.5.1
gunicorn==20.1.0
django-redis==5.2.0
Pillow==9.2.0
django-filter==23.1
drf-yasg==1.20.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.3
boto3==1.24.20

